Don't force a invoke method signarture for commands of TOption
Instead of Command{TOptions} and RootCommand{TOptions} being abstract
classes that requires a specific invoke method signature, they now
accept the name of the method to invoke and generates a command handler
based on that. This way, it is possible to add more parameters to the
invoke method, which will be resolved from the service provider.